Deleted old version of detect
